Reorganise SNS module
