+++ tests/unit/test_sqs.py
@@ -1,5 +1,7 @@
 import pytest
 
+import localstack.services.sqs.exceptions
+import localstack.services.sqs.models
 from localstack.aws.api.sqs import Message
 from localstack.services.sqs import provider
 from localstack.services.sqs.constants import DEFAULT_MAXIMUM_MESSAGE_SIZE
@@ -45,15 +47,15 @@ def test_handle_string_max_receive_count_in_dead_letter_check():
     # fmt: off
     policy = {"RedrivePolicy": "{\"deadLetterTargetArn\":\"arn:aws:sqs:us-east-1:000000000000:DeadLetterQueue\",\"maxReceiveCount\": \"5\" }"}
     # fmt: on
-    queue = provider.SqsQueue("TestQueue", "us-east-1", "123456789", policy)
-    sqs_message = provider.SqsMessage(Message(), {})
+    queue = localstack.services.sqs.models.SqsQueue("TestQueue", "us-east-1", "123456789", policy)
+    sqs_message = localstack.services.sqs.models.SqsMessage(Message(), {})
     result = provider.SqsProvider()._dead_letter_check(queue, sqs_message, None)
     assert result is False
 
 
 def test_except_check_message_size():
     message_body = "".join(("a" for _ in range(DEFAULT_MAXIMUM_MESSAGE_SIZE + 1)))
-    with pytest.raises(provider.InvalidParameterValue):
+    with pytest.raises(localstack.services.sqs.exceptions.InvalidParameterValue):
         provider.check_message_size(message_body, DEFAULT_MAXIMUM_MESSAGE_SIZE)
